+++ src/pytorch_lightning/accelerators/cpu.py
@@ -14,12 +14,12 @@
 from typing import Any, Dict, List, Union
 
 import torch
+from lightning_utilities.core.imports import RequirementCache
 
 from lightning_fabric.accelerators.cpu import _parse_cpu_cores
 from lightning_fabric.utilities.types import _DEVICE
 from pytorch_lightning.accelerators.accelerator import Accelerator
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
-from pytorch_lightning.utilities.imports import _PSUTIL_AVAILABLE
 
 
 class CPUAccelerator(Accelerator):
@@ -76,13 +76,13 @@ class CPUAccelerator(Accelerator):
 _CPU_VM_PERCENT = "cpu_vm_percent"
 _CPU_PERCENT = "cpu_percent"
 _CPU_SWAP_PERCENT = "cpu_swap_percent"
+_PSUTIL_AVAILABLE = RequirementCache("psutil")
 
 
 def get_cpu_stats() -> Dict[str, float]:
     if not _PSUTIL_AVAILABLE:
         raise ModuleNotFoundError(
-            "Fetching CPU device stats requires `psutil` to be installed."
-            " Install it by running `pip install -U psutil`."
+            f"Fetching CPU device stats requires `psutil` to be installed. {str(_PSUTIL_AVAILABLE)}"
         )
     import psutil