+++ src/lightning_app/core/app.py
@@ -105,18 +105,6 @@ class LightningApp:
                 For instance, if you want to run your app at `https://customdomain.com/myapp`,
                 set `root_path` to `/myapp`.
                 You can learn more about proxy `here <https://www.fortinet.com/resources/cyberglossary/proxy-server>`_.
-
-
-        Example:
-
-            >>> from lightning_app import LightningFlow, LightningApp
-            >>> from lightning_app.runners import MultiProcessRuntime
-            >>> class RootFlow(LightningFlow):
-            ...     def run(self):
-            ...         self.stop()
-            ...
-            >>> app = LightningApp(RootFlow())  # application can be dispatched using the `runners`.
-            >>> MultiProcessRuntime(app).dispatch()
         """
 
         self.root_path = root_path  # when running behind a proxy